# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: IoT API - Benjamin Dekeyser & Gabriel Curinga
# during dev, should point to your local machine
host: localhost:8081
# basePath prefixes all resource paths
basePath: /api
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /generic/{entity}/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity
    get:
      description: Returns an entity datas
      parameters:
        - name: entity
          type: string
          in: path
          required: true
        - name: id
          type: string
          in: path
          required: true
      # used as the method name of the controller
      operationId: getEntityById
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BasicResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /generic/{entity}:
    # binds a127 app logic to a route
    x-swagger-router-controller: entity
    get:
      description: Returns the entity datas
      parameters:
        - name: entity
          type: string
          in: path
          required: true
      # used as the method name of the controller
      operationId: getEntity
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BasicResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Insert entity
      parameters:
        - name: entity
          type: string
          in: path
          required: true
        - name: body
          in: body
          required: true
          schema:
            type: object
      # used as the method name of the controller
      operationId: insertEntity
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/BasicResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  BasicResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string